Split runner shell config tests
